+++ sd2/__init__.py
@@ -1,3 +1,3 @@
 from ._version import __version__, __version_info__
-from . import detonations, shocks, states, tools
-__all__ = ['calculate_error', 'detonations', 'shocks', 'states', 'tools']
+from . import detonations, shocks, states
+__all__ = ['calculate_error', 'detonations', 'shocks', 'states']