+++ sd2/detonations.py
@@ -11,15 +11,17 @@ import numpy as np
 import cantera as ct
 import multiprocessing as mp
 from scipy.optimize import curve_fit
-from . import tools, calculate_error, states
+from . import calculate_error, states
 
 
-def calculate_cj_state(working_gas,
-                       initial_state_gas,
-                       error_tol_temperature,
-                       error_tol_specific_volume,
-                       density_ratio,
-                       max_iterations=500):
+def calculate_cj_state(
+        working_gas,
+        initial_state_gas,
+        error_tol_temperature,
+        error_tol_velocity,
+        density_ratio,
+        max_iterations=500
+):
     """
     This function calculates the Chapman-Jouguet state and wave speed using
     Reynolds' iterative method.
@@ -35,10 +37,10 @@ def calculate_cj_state(working_gas,
         undetonated state.
     error_tol_temperature : float
         Temperature error tolerance for iteration.
-    error_tol_specific_volume : float
-        Specific volume error tolerance for iteration.
+    error_tol_velocity : float
+        Velocity error tolerance for iteration.
     density_ratio : float
-        density ratio (???).
+        density ratio.
     max_iterations : int
         Maximum number of loop iterations used to calculate output. Default
         is 500.
@@ -50,144 +52,121 @@ def calculate_cj_state(working_gas,
     initial_velocity : float
         Initial velocity resulting in the input density ratio, in m/s.
     """
-    max_iterations = int(max_iterations)
+    # initial state
+    initial_volume = 1 / initial_state_gas.density
 
-    # Set initial state
-    initial = {
-        'density': initial_state_gas.density,
-        'volume': 1 / initial_state_gas.density,
-        'pressure': initial_state_gas.P
-        }
+    # set guess values
+    guess_temperature = 2000
+    guess_velocity = 2000
+    guess_volume = initial_volume / density_ratio
+    guess_density = 1 / guess_volume
 
-    # Set initial delta values
-    delta = {
-        'temperature': 1000.,
-        'volume': 1000.,
-        'pressure': 1000.
-        }
+    # set deltas
+    delta_temperature = 1000
+    delta_velocity = 1000
 
-    # Set guess values
-    guess = {
-        'temperature': 2000.,
-        'volume': initial['volume'] / density_ratio,
-        'density': density_ratio / initial['volume']
-        }
-
-    # Add pressure and enthalpy to guess dictionary
-    guess.update(states.get_equilibrium_properties(
+    # equilibrate
+    states.get_equilibrium_properties(
         working_gas,
-        guess['density'],
-        guess['temperature'])
+        guess_density,
+        guess_temperature
     )
 
     loop_counter = 0
-    while ((abs(delta['temperature']) >
-            error_tol_temperature * guess['temperature'])
-           or
-           (abs(delta['volume']) >
-            error_tol_specific_volume * guess['volume'])):
-
-        # Manage loop count
+    while (
+            abs(delta_temperature) > (error_tol_temperature * guess_temperature)
+            or
+            abs(delta_velocity) > (error_tol_velocity * guess_velocity)
+    ):
         loop_counter += 1
+        # check for non-convergence
         if loop_counter == max_iterations:
-            warnings.warn('No convergence within {0} iterations'.
-                          format(max_iterations))
-            return [working_gas, guess['volume']]
-
-        # Calculate pressure and enthalpy error for current guess values as
-        # a numpy array, which will be the negative of the ordinate vector, b,
-        # for the linear equation Ax = b
-        ordinate_vector = (
-            calculate_error.
-            equilibrium(
-                working_gas,
-                initial_state_gas,
-                guess['volume']) * -1
+            warnings.warn(
+                'No convergence within {0} iterations'.format(max_iterations),
+                Warning
             )
+            return [working_gas, guess_velocity]
 
-        # Perturb temperature guess and find corresponding pressure and
-        # enthalpy values
-        delta['temperature'] = guess['temperature'] * 0.02
-        perturbed = tools.perturb('temperature', guess, delta)
-        perturbed.update(
-            states.get_equilibrium_properties(
-                working_gas,
-                perturbed['density'],
-                perturbed['temperature']
-                )
-            )
+        # calculate unperturbed enthalpy and pressure error for current guess
+        [error_enthalpy,
+         error_pressure] = calculate_error.equilibrium(
+            working_gas,
+            initial_state_gas,
+            guess_velocity
+        )
 
-        # Calculate pressure and enthalpy error for temperature-perturbed
-        # state, add (negative) ordinate vector to get error deltas for
-        # enthalpy and pressure, and divide by temperature delta to get rates
-        # WRT temperature
-        temperature_partials = (
-            calculate_error.
-            equilibrium(
-                working_gas,
-                initial_state_gas,
-                perturbed['volume']
-                ) +
-            ordinate_vector
-            )
-        temperature_partials /= delta['temperature']
+        # perturb temperature
+        delta_temperature = 0.02 * guess_temperature
+        perturbed_temperature = guess_temperature + delta_temperature
+        states.get_equilibrium_properties(
+            working_gas,
+            guess_density,
+            perturbed_temperature
+        )
 
-        # Perturb volume guess
-        delta['volume'] = guess['volume'] * 0.02
-        perturbed = tools.perturb('volume', guess, delta)
-        perturbed.update(
-            states.get_equilibrium_properties(
-                working_gas,
-                perturbed['density'],
-                perturbed['temperature']
-                )
-            )
+        # calculate error rates for temperature perturbed state
+        [error_perturbed_enthalpy,
+         error_perturbed_pressure] = calculate_error.equilibrium(
+            working_gas,
+            initial_state_gas,
+            guess_velocity
+        )
+        derivative_enthalpy_temperature = (error_perturbed_enthalpy -
+                                           error_enthalpy) / delta_temperature
+        derivative_pressure_temperature = (error_perturbed_pressure -
+                                           error_pressure) / delta_temperature
 
-        # Calculate pressure and enthalpy error for volume-perturbed
-        # state, add (negative) ordinate vector to get error deltas for
-        # enthalpy and pressure, and divide by volume delta to get rates
-        # WRT volume
-        volume_partials = (
-            calculate_error.
-            equilibrium(
-                working_gas,
-                initial_state_gas,
-                perturbed['volume']
-                ) +
-            ordinate_vector
-            )
-        volume_partials /= delta['volume']
+        # perturb velocity
+        delta_velocity = 0.02 * guess_velocity
+        perturbed_velocity = guess_velocity + delta_velocity
+        perturbed_temperature = guess_temperature
+        states.get_equilibrium_properties(
+            working_gas,
+            guess_density,
+            perturbed_temperature
+        )
 
-        # build coefficient matrix, A, for Ax = b
-        coefficient_matrix = np.array([temperature_partials,
-                                       volume_partials]).transpose()
+        # calculate error rates for velocity perturbed state
+        [error_perturbed_enthalpy,
+         error_perturbed_pressure] = calculate_error.equilibrium(
+            working_gas,
+            initial_state_gas,
+            perturbed_velocity
+        )
+        derivative_enthalpy_velocity = (error_perturbed_enthalpy -
+                                        error_enthalpy) / delta_velocity
+        derivative_pressure_velocity = (error_perturbed_pressure -
+                                        error_pressure) / delta_velocity
 
-        # Solve Ax = b
-        [delta['temperature'],
-         delta['volume']] = np.linalg.solve(
-             coefficient_matrix,
-             ordinate_vector
-             )
+        # invert matrix
+        jacobian = derivative_enthalpy_temperature *\
+            derivative_pressure_velocity - \
+            derivative_pressure_temperature * \
+            derivative_enthalpy_velocity
+        b = [derivative_pressure_velocity,
+             -derivative_enthalpy_velocity,
+             -derivative_pressure_temperature,
+             derivative_enthalpy_temperature]
+        a = [-error_enthalpy,
+             -error_pressure]
+        delta_temperature = (b[0] * a[0] + b[1] * a[1]) / jacobian
+        delta_velocity = (b[2] * a[0] + b[3] * a[1]) / jacobian
 
-        # Limit temperature delta to 20% of current guess
-        max_temperature_delta = guess['temperature'] * 0.2
-        if abs(delta['temperature']) > max_temperature_delta:
-            delta['temperature'] *= (
-                max_temperature_delta / abs(delta['temperature'])
-                )
+        # limit temperature changes
+        max_temperature_delta = 0.2 * guess_temperature
+        if abs(delta_temperature) > max_temperature_delta:
+            delta_temperature *= max_temperature_delta / abs(delta_temperature)
 
-        # Update guess values
-        guess['temperature'] += delta['temperature']
-        guess['volume'] += delta['volume']
-        guess.update(
-            states.get_equilibrium_properties(
-                working_gas,
-                guess['density'],
-                guess['temperature']
-                )
-            )
+        # apply deltas and equilibrate
+        guess_temperature += delta_temperature
+        guess_velocity += delta_velocity
+        states.get_equilibrium_properties(
+            working_gas,
+            guess_density,
+            guess_temperature
+        )
 
-    return [working_gas, guess['volume']]
+    return [working_gas, guess_velocity]
 
 
 def _calculate_over_ratio_range(
@@ -198,7 +177,7 @@ def _calculate_over_ratio_range(
         species_mole_fractions,
         mechanism,
         error_tol_temperature,
-        error_tol_specific_volume
+        error_tol_velocity
 ):
     initial_state_gas = ct.Solution(mechanism)
     initial_state_gas.TPX = [
@@ -217,19 +196,22 @@ def _calculate_over_ratio_range(
         working_gas,
         initial_state_gas,
         error_tol_temperature,
-        error_tol_specific_volume,
+        error_tol_velocity,
         current_density_ratio
     )
 
     return current_state_number, current_velocity
 
 
-def calculate_cj_speed(initial_pressure,
-                       initial_temperature,
-                       species_mole_fractions,
-                       mechanism,
-                       return_r_squared=False,
-                       return_state=False):
+def calculate_cj_speed(
+        initial_pressure,
+        initial_temperature,
+        species_mole_fractions,
+        mechanism,
+        use_multiprocessing=False,
+        return_r_squared=False,
+        return_state=False
+):
     """
     This function calculates CJ detonation velocity
 
@@ -245,6 +227,8 @@ def calculate_cj_speed(initial_pressure,
         string or dictionary of reactant species mole fractions
     mechanism : str
         cti file containing mechanism data (e.g. 'gri30.cti')
+    use_multiprocessing : bool
+        use multiprocessing to speed up CJ speed calculation
     return_r_squared : bool
         return the R^2 value of the CJ speed vs. density ratio fit
     return_state : bool
@@ -258,13 +242,13 @@ def calculate_cj_speed(initial_pressure,
     num_steps = 5
     max_density_ratio = 2.0
     min_density_ratio = 1.5
-    pool = mp.Pool()
 
-    # TODO: figure out what to do about CJ state calculation
+    if use_multiprocessing:
+        pool = mp.Pool()
 
     # Set error tolerances for CJ state calculation
     error_tol_temperature = 1e-4
-    error_tol_specific_volume = 1e-4
+    error_tol_velocity = 1e-4
 
     counter = 1
     r_squared = 0.0
@@ -285,23 +269,40 @@ def calculate_cj_speed(initial_pressure,
             num_steps + 1
         )
 
-        # parallel loop through density ratios
-        stargs = [[number,
-                   ratio,
-                   initial_temperature,
-                   initial_pressure,
-                   species_mole_fractions,
-                   mechanism,
-                   error_tol_temperature,
-                   error_tol_specific_volume
-                   ]
-                  for number, ratio in
-                  zip(
-                      range(len(density_ratio_array)),
-                      density_ratio_array
-                  )
-                  ]
-        result = pool.starmap(_calculate_over_ratio_range, stargs)
+        if use_multiprocessing:
+            # parallel loop through density ratios
+            stargs = [[number,
+                       ratio,
+                       initial_temperature,
+                       initial_pressure,
+                       species_mole_fractions,
+                       mechanism,
+                       error_tol_temperature,
+                       error_tol_velocity
+                       ]
+                      for number, ratio in
+                      zip(
+                          range(len(density_ratio_array)),
+                          density_ratio_array
+                      )
+                      ]
+            # noinspection PyUnboundLocalVariable
+            result = pool.starmap(_calculate_over_ratio_range, stargs)
+
+        else:
+            # no multiprocessing, just use map
+            result = list(map(
+                _calculate_over_ratio_range,
+                [item for item in range(len(density_ratio_array))],
+                density_ratio_array,
+                [initial_temperature for _ in density_ratio_array],
+                [initial_pressure for _ in density_ratio_array],
+                [species_mole_fractions for _ in density_ratio_array],
+                [mechanism for _ in density_ratio_array],
+                [error_tol_temperature for _ in density_ratio_array],
+                [error_tol_velocity for _ in density_ratio_array]
+            ))
+
         result.sort()
         cj_velocity_calculations = np.array(
             [item for (_, item) in result]
@@ -339,6 +340,7 @@ def calculate_cj_speed(initial_pressure,
         max_density_ratio = adjusted_density_ratio * (1 + 0.001)
         counter += 1
 
+    # noinspection PyUnboundLocalVariable
     cj_speed = curve_fit_function(
         adjusted_density_ratio,
         *curve_fit_coefficients
@@ -361,16 +363,18 @@ def calculate_cj_speed(initial_pressure,
             working_gas,
             initial_state_gas,
             error_tol_temperature,
-            error_tol_specific_volume,
+            error_tol_velocity,
             adjusted_density_ratio
         )
 
     if return_r_squared and return_state:
+        # noinspection PyUnboundLocalVariable
         return {'cj speed': cj_speed,
                 'R^2': r_squared,
                 'cj state': working_gas
                 }
     elif return_state:
+        # noinspection PyUnboundLocalVariable
         return {'cj speed': cj_speed,
                 'cj state': working_gas
                 }