+++ sd2/states.py
@@ -7,7 +7,11 @@ http://www.galcit.caltech.edu/EDL/public/cantera/html/SD_Toolbox/
 """
 
 
-def get_equilibrium_properties(gas, density, temperature):
+def get_equilibrium_properties(
+        gas,
+        density,
+        temperature
+):
     """
     This function calculates the equilibrium pressure and enthalpy given
     temperature and density
@@ -37,7 +41,11 @@ def get_equilibrium_properties(gas, density, temperature):
     return {'pressure': pressure, 'enthalpy': enthalpy}
 
 
-def get_frozen_properties(gas, density, temperature):
+def get_frozen_properties(
+        gas,
+        density,
+        temperature
+):
     """
     This function calculates the frozen pressure and enthalpy given temperature
     and density